// src/model/paciente.ts
import { Sociedad } from "./sociedad";

export class Paciente {
    ciudad: string;
    correoElectronico: string;
    direccion: string;
    edad: number;
    estatus: string;
    fechaCreacion: string;
    fechaModificacion: string;
    idPaciente: number;
    nombre: string;
    ocupacion: string;
    sexo: string;
    sociedad: Sociedad;
    telefono: string;
}